+++ esp-hal/src/soc/esp32s3/peripherals.rs
@@ -72,6 +72,7 @@ crate::peripherals! {
         SW_INTERRUPT <= virtual,
         TIMG0 <= TIMG0,
         TIMG1 <= TIMG1,
+        TSENS <= virtual,
         TWAI0 <= TWAI0,
         UHCI0 <= UHCI0,
         UHCI1 <= UHCI1,