+++ qa-test/src/bin/temperature_sensor.rs
@@ -1,7 +1,7 @@
 //! This example uses the internal temperature sensor to measure the chip's
 //! temperature
 
-//% CHIPS: esp32c6 esp32c3
+//% CHIPS: esp32c6 esp32c3 esp32s3
 
 #![no_std]
 #![no_main]